Move rustc_parse diagnostic structs to separate module
